Remove log error when DOI is not set and rm tryes to rm the empty file

+++ models/WorkflowUpload.php
@@ -197,7 +197,10 @@ class WorkflowUpload extends \yii\db\ActiveRecord
         {
             case 0:
                 $success="Workflow successfully uploaded!";
-                Software::exec_log("rm $doiFile");
+                if($doiFile!='\'\'')
+                {
+                    Software::exec_log("rm $doiFile");
+                }
                 break;
             case 2:
                 $errors.="Error: code $ret. ";
